Reworked timer a bit

+++ include/timer.h
@@ -4,14 +4,14 @@
 
 struct timer
 {
-  double delay_ms;
+  unsigned long delay_ms;
   void (*callback)(void);
 };
 
 /**
  * @brief Adds a timer to run the callback after delay_ms usecs. Runs in separate thread.
- * 
- * @param delay_ms 
- * @param callback 
+ *
+ * @param delay_ms
+ * @param callback
  */
-void add_timer(double delay_ms, void (*callback)(void));
+void add_timer(unsigned long delay_ms, void (*callback)(void));
